Login and getting projects are working on server
Login and getting projects are working on the server.
Currently server.js is calling the functions directly as a test.
I will need to implement actual routing, store the authentication
token somewhere so that I don't have to login before every request,
and redirect the user if they try to request projects before
logging in.